requirements.txt
asgiref==3.2.10
colour==0.1.5
cycler==0.10.0
Django==3.0.8
djangorestframework==3.11.0
kiwisolver==1.2.0
matplotlib==3.3.0
numpy==1.19.1
Pillow==7.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
